requirements.txt
python-dateutil==2.8.1
requests==2.24.0
retry==0.9.2
Pybooru==4.2.2
